requirements.txt
fastapi[all]
openai
pydantic
python-dotenv
pinecone
